[#930] Reorganising packages
