+++ app/src/androidTest/java/org/akvo/flow/activity/form/data/SurveyInstaller.java
@@ -17,7 +17,7 @@
  *  along with Akvo Flow.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.akvo.flow.activity.testhelper;
+package org.akvo.flow.activity.form.data;
 
 import android.content.Context;
 import android.text.TextUtils;
