+++ app/src/androidTest/java/org/akvo/flow/activity/form/views/DateQuestionViewTest.java
@@ -18,7 +18,7 @@
  *
  */
 
-package org.akvo.flow.activity;
+package org.akvo.flow.activity.form.views;
 
 import android.content.Context;
 import android.content.Intent;
@@ -30,8 +30,10 @@ import android.support.test.runner.AndroidJUnit4;
 import android.widget.DatePicker;
 
 import org.akvo.flow.R;
-import org.akvo.flow.activity.testhelper.SurveyInstaller;
-import org.akvo.flow.activity.testhelper.SurveyRequisite;
+import org.akvo.flow.activity.Constants;
+import org.akvo.flow.activity.FormActivity;
+import org.akvo.flow.activity.form.data.SurveyInstaller;
+import org.akvo.flow.activity.form.data.SurveyRequisite;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
@@ -49,16 +51,16 @@ import static android.support.test.espresso.matcher.ViewMatchers.withClassName;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
 import static java.util.Calendar.SHORT;
-import static org.akvo.flow.activity.FormActivityTestUtil.clickNext;
-import static org.akvo.flow.activity.FormActivityTestUtil.getFormActivityIntent;
-import static org.akvo.flow.activity.FormActivityTestUtil.verifySubmitButtonDisabled;
-import static org.akvo.flow.activity.FormActivityTestUtil.verifySubmitButtonEnabled;
+import static org.akvo.flow.activity.form.FormActivityTestUtil.clickNext;
+import static org.akvo.flow.activity.form.FormActivityTestUtil.getFormActivityIntent;
+import static org.akvo.flow.activity.form.FormActivityTestUtil.verifySubmitButtonDisabled;
+import static org.akvo.flow.activity.form.FormActivityTestUtil.verifySubmitButtonEnabled;
 import static org.akvo.flow.tests.R.raw.datesurvey;
 import static org.hamcrest.Matchers.endsWith;
 
 @MediumTest
 @RunWith(AndroidJUnit4.class)
-public class DateFormActivityTest {
+public class DateQuestionViewTest {
 
     private static SurveyInstaller installer;
