+++ app/src/androidTest/java/org/akvo/flow/activity/form/views/OptionsQuestionViewTest.java
@@ -1,23 +1,24 @@
 /*
- *  Copyright (C) 2017 Stichting Akvo (Akvo Foundation)
+ * Copyright (C) 2017 Stichting Akvo (Akvo Foundation)
  *
- *  This file is part of Akvo Flow.
+ * This file is part of Akvo Flow.
  *
- *  Akvo Flow is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
+ * Akvo Flow is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
  *
- *  Akvo Flow is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
+ * Akvo Flow is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
  *
- *  You should have received a copy of the GNU General Public License
- *  along with Akvo Flow.  If not, see <http://www.gnu.org/licenses/>.
+ * You should have received a copy of the GNU General Public License
+ * along with Akvo Flow.  If not, see <http://www.gnu.org/licenses/>.
+ *
  */
 
-package org.akvo.flow.activity;
+package org.akvo.flow.activity.form.views;
 
 import android.content.Context;
 import android.content.Intent;
@@ -26,8 +27,9 @@ import android.support.test.filters.MediumTest;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 
-import org.akvo.flow.activity.testhelper.SurveyInstaller;
-import org.akvo.flow.activity.testhelper.SurveyRequisite;
+import org.akvo.flow.activity.FormActivity;
+import org.akvo.flow.activity.form.data.SurveyInstaller;
+import org.akvo.flow.activity.form.data.SurveyRequisite;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
@@ -42,17 +44,17 @@ import static android.support.test.espresso.matcher.ViewMatchers.isChecked;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static org.akvo.flow.activity.Constants.TEST_FORM_SURVEY_INSTANCE_ID;
-import static org.akvo.flow.activity.FormActivityTestUtil.clickNext;
-import static org.akvo.flow.activity.FormActivityTestUtil.getFormActivityIntent;
-import static org.akvo.flow.activity.FormActivityTestUtil.matchToolbarTitle;
-import static org.akvo.flow.activity.FormActivityTestUtil.verifyQuestionTitleDisplayed;
-import static org.akvo.flow.activity.FormActivityTestUtil.verifySubmitButtonDisabled;
-import static org.akvo.flow.activity.FormActivityTestUtil.verifySubmitButtonEnabled;
+import static org.akvo.flow.activity.form.FormActivityTestUtil.clickNext;
+import static org.akvo.flow.activity.form.FormActivityTestUtil.getFormActivityIntent;
+import static org.akvo.flow.activity.form.FormActivityTestUtil.matchToolbarTitle;
+import static org.akvo.flow.activity.form.FormActivityTestUtil.verifyQuestionTitleDisplayed;
+import static org.akvo.flow.activity.form.FormActivityTestUtil.verifySubmitButtonDisabled;
+import static org.akvo.flow.activity.form.FormActivityTestUtil.verifySubmitButtonEnabled;
 import static org.akvo.flow.tests.R.raw.optionsurvey;
 
 @MediumTest
 @RunWith(AndroidJUnit4.class)
-public class OptionsFormActivityTest {
+public class OptionsQuestionViewTest {
 
     private static final String FORM_TITLE = "OptionsQuestionForm";
     private static SurveyInstaller installer;
